수정
1. GourmetType Default 설정
2. 인증 이메일 HTML 코드 수정
3. CICD Redis 서버 재부팅 설정

+++ domain-user/src/main/java/co/dalicious/domain/user/converter/GourmetTypeConverter.java
@@ -6,14 +6,14 @@ import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
 @Converter
-public class GourmetTypeConverter implements AttributeConverter<GourmetType, Long> {
+public class GourmetTypeConverter implements AttributeConverter<GourmetType, Integer> {
     @Override
-    public Long convertToDatabaseColumn(GourmetType gourmetType) {
+    public Integer convertToDatabaseColumn(GourmetType gourmetType) {
         return gourmetType.getCode();
     }
 
     @Override
-    public GourmetType convertToEntityAttribute(Long dbData) {
+    public GourmetType convertToEntityAttribute(Integer dbData) {
         return GourmetType.ofCode(dbData);
     }
 }
